Fixing error when creating redirect session (#59)
The function `get_access_token` fetched the wrong setting, `client_id` instead of `production_client_id`. When being used in session initialization, this function returned an empty access token, and we're unable to create the session.

I think this function has never worked, and it just happened to not being called.